ai-poc/services/indexer/requirements.txt
pypdf
python-docx
requests

